add crud preparation functions to columnDef/init.js
